Spectrum retrieval now works (for MSSM only, but can be expanded as needed). See spectrum_retrive_demo.yaml for example usage.
Note: R1 datasets are missing the "scale(Q)" output, which gives the scale
at which DRbar parameters are defined. However, we know that this is MSUSY
for all those scans, so I added a fallback to compute MSUSY from the other
DRbar parameters as sqrt(mst1*mst2), where the stop masses are the DRbar
masses (which aren't directly in the output but can be computed from other
stuff). This agrees with FlexibleSUSY reporting of MSUSY to 5 sig figures,
but I know no reason why it shouldn't be exact. So could be a bug; will
continue testing.

	modified:   Printers/src/printers/hdf5printer/retrieve_overloads.cpp
	modified:   SpecBit/src/SpecBit_MSSM.cpp
	modified:   yaml_files/spectrum_retrieve_demo.yaml